+++ packages/notification-services-controller/src/NotificationServicesController/mocks/index.ts
@@ -1,6 +1,4 @@
 export * from './mock-feature-announcements';
-export * from './mock-notification-trigger';
-export * from './mock-notification-user-storage';
 export * from './mock-raw-notifications';
 export * from './mockResponses';
 export * from './mock-snap-notification';