+++ packages/profile-sync-controller/src/controllers/user-storage/__fixtures__/mockMessenger.ts
@@ -75,6 +75,8 @@ export function createCustomUserStorageMessenger(props?: {
       'AccountsController:accountRenamed',
       'AccountsController:accountAdded',
       'NetworkController:networkRemoved',
+      'AddressBookController:contactUpdated',
+      'AddressBookController:contactDeleted',
     ],
   });